+++ src/App.tsx
@@ -1,8 +1,8 @@
-import Home from "./pages/Home";
+import HomePage from "./pages/HomePage";
 
 function App() {
     return ( 
-        <Home/>
+        <HomePage/>
      );
 }
